refactor(auth): file structure
