+++ api/actix/src/apps/auth/mod.rs
@@ -1,0 +1,4 @@
+pub mod models;
+pub mod urls;
+pub mod views;
+pub use urls::url;
