+++ api/actix/src/apps/auth/models.rs
@@ -1,0 +1,12 @@
+use serde::Deserialize;
+use validator::Validate;
+
+#[derive(Debug, Deserialize, Validate)]
+pub struct NewUser {
+    #[validate(length(min = 5))]
+    pub username: String,
+    #[validate(email)]
+    pub email: String,
+    #[validate(length(min = 8))]
+    pub password: String,
+}
