+++ api/actix/src/apps/auth/views.rs
@@ -1,23 +1,20 @@
 use std::ops::Deref;
 
 use actix_web::{
-    get, post,
     web::{Data, Json},
     HttpResponse, Responder,
 };
 use actix_web_httpauth::extractors::basic::BasicAuth;
-use api_models::{
-    user::{CreateUserBody, UserNoPassword},
-    AppState, TokenClaims,
-};
 use argonautica::{Hasher, Verifier};
 use chrono::Utc;
 use hmac::{Hmac, Mac};
 use jwt::SignWithKey;
 use sha2::Sha512;
 
-#[post("/regiseter")]
-pub async fn register(state: Data<AppState>, body: Json<CreateUserBody>) -> impl Responder {
+use super::models::NewUser;
+use crate::prelude::{AppState, TokenClaims};
+
+pub async fn register(state: Data<AppState>, body: Json<NewUser>) -> impl Responder {
     let user = body.into_inner();
 
     let mut hasher = Hasher::default();
@@ -27,28 +24,25 @@ pub async fn register(state: Data<AppState>, body: Json<CreateUserBody>) -> impl
         .hash()
         .unwrap();
 
-    let query = sqlx::query_as!(
-        UserNoPassword,
+    let query = sqlx::query!(
         r#"
-        INSERT INTO "User" (username, first_name, last_name, email, password)
-          VALUES ($1, $2, $3, $4, $5) RETURNING id, username
+        INSERT INTO "User" (username, email, password, last_login)
+          VALUES ($1, $2, $3, $4) RETURNING id
         "#,
         user.username,
-        user.first_name,
-        user.last_name,
         user.email,
-        hashed_password
+        hashed_password,
+        Utc::now().naive_utc(),
     )
     .fetch_one(&state.pool)
     .await;
 
     match query {
-        Ok(user) => HttpResponse::Ok().json(user),
+        Ok(user) => HttpResponse::Ok().json(format!("{}", user.id)),
         Err(e) => HttpResponse::InternalServerError().json(format!("{:?}", e)),
     }
 }
 
-#[get("/login")]
 pub async fn login(state: Data<AppState>, credentials: BasicAuth) -> impl Responder {
     let jwt_secret: Hmac<Sha512> = Hmac::new_from_slice(shared::jwt_secret().deref()).unwrap();
     let username = credentials.user_id();
@@ -103,4 +97,8 @@ pub async fn login(state: Data<AppState>, credentials: BasicAuth) -> impl Respon
         }
         None => HttpResponse::Unauthorized().json("Must provide username and password"),
     }
+}
+
+pub async fn logout(state: Data<AppState>) -> impl Responder {
+    format!("")
 }
