+++ api/actix/src/main.rs
@@ -1,8 +1,9 @@
-mod routes;
+mod apps;
+mod prelude;
 
 use actix_web::{web::Data, App, HttpServer};
-use api_models::{pool, AppState};
-use routes::auth::{login, register};
+use apps::auth;
+use prelude::{pool, AppState};
 use std::env;
 
 #[actix_web::main]
@@ -16,13 +17,12 @@ async fn main() -> std::io::Result<()> {
     base_path.push("migrations");
 
     let migrator = sqlx::migrate::Migrator::new(base_path).await.unwrap();
-    migrator.run(&pool).await.unwrap();
+    migrator.run(&pool).await.expect("Could not run migrations");
 
     HttpServer::new(move || {
         App::new()
             .app_data(Data::new(AppState::new(pool.to_owned())))
-            .service(login)
-            .service(register)
+            .service(auth::url())
     })
     .bind(("127.0.0.1", 8080))?
     .run()
