--- api/models/Cargo.toml
@@ -1,19 +1,0 @@
-[package]
-name = "api-models"
-version = "0.1.0"
-edition = "2021"
-publish = false
-
-[dependencies]
-api-lib = { workspace = true }
-shared = { workspace = true }
-sqlx = { workspace = true }
-tracing = { workspace = true }
-tracing-subscriber = { version = "0.3", features = [
-  "env-filter",
-  "json",
-  "time",
-] }
-color-eyre = { workspace = true }
-chrono = { workspace = true }
-serde = { workspace = true }
